module4-task1 CR Fixes
