--- src/components/offers-list/offers-list.tsx
@@ -1,55 +1,0 @@
-// import { useState } from 'react';
-import { OfferPreview } from '../../types/offer';
-import { getWordEndingByCount } from '../../utils/common';
-import PlaceCard from '../place-card/place-card';
-import { CityName } from '../../const';
-
-type OffersListProps = {
-  offers: OfferPreview[];
-};
-
-function OffersList({ offers }: OffersListProps): JSX.Element {
-  // const activeOfferId = useState<OfferPreview['id'] | null>(null);
-  // пригодится позже
-  const activeCityName = CityName.Amsterdam;
-
-  return (
-    <div className="cities">
-      <div className="cities__places-container container">
-        <section className="cities__places places">
-          <h2 className="visually-hidden">Places</h2>
-          <b className="places__found">{offers.length} place{getWordEndingByCount(offers.length)} to stay in {activeCityName}</b>
-          <form className="places__sorting" action="#" method="get">
-            <span className="places__sorting-caption">Sort by</span>
-            <span className="places__sorting-type" tabIndex={0}>
-              Popular
-              <svg className="places__sorting-arrow" width="7" height="4">
-                <use xlinkHref="#icon-arrow-select" />
-              </svg>
-            </span>
-            <ul className="places__options places__options--custom places__options--opened">
-              <li className="places__option places__option--active" tabIndex={0}>Popular</li>
-              <li className="places__option" tabIndex={0}>Price: low to high</li>
-              <li className="places__option" tabIndex={0}>Price: high to low</li>
-              <li className="places__option" tabIndex={0}>Top rated first</li>
-            </ul>
-          </form>
-          <div className="cities__places-list places__list tabs__content">
-            {offers.map((offer) => (
-              <PlaceCard
-                key={offer.id}
-                offer={offer}
-                block="cities"
-              />
-            ))}
-          </div>
-        </section>
-        <div className="cities__right-section">
-          <section className="cities__map map"></section>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-export default OffersList;
